fix(maple-cloud): SearchAsset.mtime accepts fractional epoch ms
User-visible symptom: 2026-04 timeline section showed no images on
the iPhone even after the page=0 fix. Logs:

  decode SearchResponse failed (library …, 2026-4):
    "mtime":1776035930475.9543

JSONDecoder bails on the whole response when one nested mtime is a
Double (e.g. a panorama asset whose MongoDB record holds a
Decimal128/Double instead of NumberLong). The 2026-04 bucket has one
such asset, so all 15 results decoded together → []. 2026-03 has no
fractional-mtime assets, so it decoded fine (76 results).

Widen \`mtime: Int64?\` → \`mtime: Double?\` to tolerate both shapes.
Integer milliseconds round-trip without precision loss; callers can
truncate if they need Int64.

+++ src/apple/Packages/MapleCore/Sources/MapleCore/Cloud/CloudSearchTypes.swift
@@ -37,8 +37,13 @@ public struct SearchAsset: Codable, Equatable, Sendable {
   public let abs_path: String
   public let filename: String
   public let size: Int64?
-  /// Last-modified epoch ms — JSON number on the wire (not a string).
-  public let mtime: Int64?
+  /// Last-modified epoch ms. Wire format is a JSON number — usually an
+  /// integer, but the server sometimes sends a fractional value (e.g.
+  /// `1776035930475.9543` for a few panorama assets where MongoDB
+  /// returned a Decimal128 / Double instead of a NumberLong). Decoding
+  /// as Double tolerates both shapes; integer milliseconds round-trip
+  /// without loss. Truncate to Int64 if the caller needs that.
+  public let mtime: Double?
   public let captured_at: String?
   public let camera: SearchAssetCamera?
   public let lens: String?
